dataset/traffic-generator/requirements.txt
psycopg2-binary
requests
numpy